fedup.grubby: add set_next_boot()
We can use this to make the system boot the given entry for the next
boot *only*. Handy.

Note that it only works for GRUB2 systems at the moment. This is
intentional; I'd prefer it if all the magic "if bootloader is zipl.."
abstraction stay in grubby, so I'd rather write a patch for that than
handle it all here.

When grubby handles that, we'll use grubby to do it. Until then...